packaging: rename supervdsmServer.py to supervdsmServer and add environment file
supervdsm process is started by "python supervdsmServer.pyc". This is a
bit verbose when it is put in the SysV init script and SystemD service
unit. This patch renames supervdsmServer.py to supervdsmServer to enable
the init script start supervdsmServer directly.

In this patch, an environment file is also add for supervdsmd service.

Change-Id: Ifbcfcfb6f22040d038541c79ac4d0e66ac106504
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>