libvirt configure: use "ed" instead of "sed --copy -i"
"sed -i" generates a temp file for output, then renames the temp file to
the targeted file. Renaming a file means just changing the content of
the directory entry, and the edited file has different inode index from
the original version. So the sed edited file is a different file though
with the same file name as the original version. To use sed on
ovirt-node on bind-mounted files, we need "--copy" to copy the temp file
into targeted file instead of renaming. Unfortunately, "--copy" is
provided by a patch that is specific to the RPM packaged sed on Red Hat
family distros. The "--copy" option is neither in GNU sed [1] nor in
Single Unix Specification sed [2].

In this patch we use "ed" instead of "sed -i --copy". "ed" edits the
file in-place, by buffering the edited contents in memory and flushing
to the file at last.

[1] http://www.gnu.org/software/sed/manual/sed.html
[2] http://pubs.opengroup.org/onlinepubs/9699919799/utilities/sed.html

Change-Id: Ic265df1a5abca0104d044f8c3048cf0643547c70
Signed-off-by: Zhou Zheng Sheng <zhshzhou@linux.vnet.ibm.com>